Docker (#29)
### Problem
We are running our app in different environments between us, and it'll be run in a different environment for production as well. This was causing some issues with versions and dependencies.

### Solution
Use Docker to create a Docker image for the `client` and `server`. Then, use `docker-compose` to create a composition of the two images. Now, we just need to run `docker-compose build` once and then `docker-compose up` each time we want to spin up the containers. And then run `docker-compose down` to spin them down when you're done working on it for that session.

### Testing
This was tested on my machine by cleaning and rebuilding the images using `docker-compose` and running the site to make sure hot reloading and nodemon were both still functional.

### Notes
You need to first install Docker onto your system and go through the setup steps before this can work.