Dashboard Widget Cards (#41)
### Problem
We needed to create a dashboard with all of the main widgets we plan on implementing.

### Solution
This PR maintains the original grid of cards idea except it uses a flexbox to make the cards responsive. The dashboard currently has room for 6 widgets which we will probably end up filling up. All of these widgets have `name`, `description`, and `icon` properties passed to them and we can eventually create custom SVGs for our icons to mimic the Google Suite look.

### Testing
This was tested by observing how cards behave at different page sizes. Generally, the cards will shrink and scale as a function of the page height and width. Text doesn't currently scale but I didn't think it was a huge problem.